Add files via upload
This is my first time uploading something here. it is supposed to be a simple proxy rotator. the code i have here is designed to find a list of valid proxies. at the time of uploading this file though, i am having troubles with the "except" statement on line 23. it is my first time using an except statement so i will need to do more research on how they need to function.